requirements.txt
Django==4.2.7
django-allauth==0.57.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
django-redis==5.4.0
Pillow==10.1.0
python-dateutil==2.8.2
jpholiday==0.1.8
psycopg2-binary==2.9.7
oracledb>=2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.4
django-celery-beat==2.5.0